Fixed a rare crash and type of scores + a Sign math utility
[FIXED]     Dismissing a layer it's already being dismissed crashed.
[UPDATED]   Scores can be 64-bit ints, like in GameKit.
[ADDED]     Sign math utility that determines the sign of a number.

+++ Pearl-Cocos2D/View/HUDLayer.h
@@ -37,7 +37,7 @@
 @property (nonatomic, readonly, retain) BarLayer        *messageBar;
 
 -(void) updateHudWasGood:(BOOL)wasGood;
--(void) updateHudWithNewScore:(int)newScore wasGood:(BOOL)wasGood;
+-(void) updateHudWithNewScore:(int64_t)newScore wasGood:(BOOL)wasGood;
 -(BOOL) hitsHud: (CGPoint)pos;
 
 @end
